move more logic to content detail workspace base
